// projects/ngx-sp-infra/src/lib/components/lib-integracao-ldap/models/InfraLDAP.ts
export class InfraLDAP 
{
    Tenant_Id: number = 0;
    Id: number = 0;
    Descricao: string = "";
    Hostname: string = "";
    Username: string = "";
    Password: string = "";
    Path: string = "";
    SearchDN: boolean = false;
    GrupodDefaultId: number = 0;
    Is_RequireAprov: boolean = false;
    Is_Active: boolean = false;
    Is_LDPAS: boolean = false;
}